+++ tests/ApiTest.php
@@ -16,6 +16,7 @@ class ApiTest extends ApiTestCase
 
     protected function setUp(): void
     {
+        ApiTestCase::$alwaysBootKernel = true;
         $this->testClient = new TestClient(ApiTestCase::createClient());
         $this->testClient->setUpUser();